adapter,compute-client: move caught up checks to adapter
@teskje noticed [0] that the caught up check would more neatly
generalize to storage collections if it lived in the coordinator instead
of the compute controller.

This commit is nearly pure code movement, but there are some small
refactorings required in order to use only the public API methods from
the compute controller.

[0]: https://github.com/MaterializeInc/materialize/pull/29699#discussion_r1771024070
