+++ src/compute-client/src/controller/instance.rs
@@ -646,101 +646,6 @@ impl<T: ComputeControllerTimestamp> Instance<T> {
         Ok(all_hydrated)
     }
 
-    /// Returns `true` if all non-transient, non-excluded collections have their write
-    /// frontier (aka. upper) within `allowed_lag` of the "live" frontier
-    /// reported in `live_frontiers`. The "live" frontiers are frontiers as
-    /// reported by a currently running `environmentd` deployment, during a 0dt
-    /// upgrade.
-    ///
-    /// Collections whose write frontier is behind `now` by more than the cutoff
-    /// are ignored.
-    ///
-    /// This also returns `true` in case this cluster does not have any
-    /// replicas.
-    pub fn collections_caught_up(
-        &self,
-        allowed_lag: T,
-        cutoff: T,
-        now: T,
-        live_frontiers: &BTreeMap<GlobalId, Antichain<T>>,
-        exclude_collections: &BTreeSet<GlobalId>,
-    ) -> bool {
-        if self.replicas.is_empty() {
-            return true;
-        }
-
-        let mut all_caught_up = true;
-
-        for (id, collection) in self.collections_iter() {
-            if id.is_transient() || exclude_collections.contains(&id) {
-                // These have no relation to dataflows running on previous
-                // deployments.
-                continue;
-            }
-
-            let write_frontier = collection.write_frontier();
-
-            let live_write_frontier = match live_frontiers.get(&id) {
-                Some(frontier) => frontier,
-                None => {
-                    // The collection didn't previously exist, so consider
-                    // ourselves hydrated as long as our write_ts is > 0.
-                    tracing::info!(?write_frontier, "collection {id} not in live frontiers");
-                    // The collection didn't previously exist, so consider
-                    // ourselves hydrated as long as our write_ts is > 0.
-                    if write_frontier.less_equal(&T::minimum()) {
-                        all_caught_up = false;
-                    }
-                    continue;
-                }
-            };
-
-            // We can't do comparisons and subtractions, so we bump up the live
-            // write frontier by the cutoff, and then compare that against
-            // `now`.
-            let live_write_frontier_plus_cutoff = live_write_frontier
-                .iter()
-                .map(|t| t.step_forward_by(&cutoff));
-            let live_write_frontier_plus_cutoff =
-                Antichain::from_iter(live_write_frontier_plus_cutoff);
-
-            let beyond_all_hope = live_write_frontier_plus_cutoff.less_equal(&now);
-
-            if beyond_all_hope {
-                tracing::info!(?live_write_frontier, ?now, "live write frontier of collection {id} is too far behind 'now', ignoring for caught-up checks");
-                continue;
-            }
-
-            // We can't do easy comparisons and subtractions, so we bump up the
-            // write frontier by the allowed lag, and then compare that against
-            // the write frontier.
-            let write_frontier_plus_allowed_lag = write_frontier
-                .iter()
-                .map(|t| t.step_forward_by(&allowed_lag));
-            let bumped_write_plus_allowed_lag =
-                Antichain::from_iter(write_frontier_plus_allowed_lag);
-
-            let within_lag =
-                PartialOrder::less_equal(live_write_frontier, &bumped_write_plus_allowed_lag);
-
-            if !within_lag {
-                // We are not within the allowed lag!
-                //
-                // We continue with our loop instead of breaking out early, so
-                // that we log all non-caught-up replicas.
-                tracing::info!(
-                    ?write_frontier,
-                    ?live_write_frontier,
-                    ?allowed_lag,
-                    "collection {id} is not caught up"
-                );
-                all_caught_up = false;
-            }
-        }
-
-        all_caught_up
-    }
-
     /// Returns `true` if all non-transient, non-excluded collections are hydrated on at least one
     /// replica.
     ///
